requirements.txt
sqlalchemy
aiosqlite

fastapi
uvicorn
authx

pydantic
pydantic-settings

pyjwt[crypto]
bcrypt

python-multipart